Added intro and game flow class and test files
