+++ lib/game_flow.rb
@@ -1,0 +1,9 @@
+
+  class GameFlow
+  attr_reader :decision
+
+
+  def quit
+  end
+
+end
